[Admin] Add NavigationTrait and use it in product pages
