+++ src/Sylius/Behat/Page/Admin/Product/UpdateConfigurableProductPage.php
@@ -16,8 +16,7 @@ namespace Sylius\Behat\Page\Admin\Product;
 use Behat\Mink\Element\NodeElement;
 use Behat\Mink\Session;
 use Sylius\Behat\Behaviour\ChecksCodeImmutability;
-use Sylius\Behat\Context\Ui\Admin\Helper\EditToShowPageSwitcherTrait;
-use Sylius\Behat\Context\Ui\Admin\Helper\ShowPageButtonCheckerTrait;
+use Sylius\Behat\Context\Ui\Admin\Helper\NavigationTrait;
 use Sylius\Behat\Page\Admin\Crud\UpdatePage as BaseUpdatePage;
 use Sylius\Behat\Service\AutocompleteHelper;
 use Sylius\Behat\Service\Helper\AutocompleteHelperInterface;
@@ -26,8 +25,7 @@ use Symfony\Component\Routing\RouterInterface;
 class UpdateConfigurableProductPage extends BaseUpdatePage implements UpdateConfigurableProductPageInterface
 {
     use ChecksCodeImmutability;
-    use EditToShowPageSwitcherTrait;
-    use ShowPageButtonCheckerTrait;
+    use NavigationTrait;
 
     /**
      * @param array<array-key, string> $minkParameters
@@ -40,13 +38,11 @@ class UpdateConfigurableProductPage extends BaseUpdatePage implements UpdateConf
         private AutocompleteHelperInterface $autocompleteHelper,
     ) {
         parent::__construct($session, $minkParameters, $router, $routeName);
-
-        $this->defineResourceName();
     }
 
-    protected function defineResourceName(): void
+    protected function getResourceName(): string
     {
-        $this->resourceName = 'product';
+        return 'product';
     }
 
     public function saveChanges(): void
