+++ src/Sylius/Behat/Page/Admin/Product/UpdateSimpleProductPage.php
@@ -16,8 +16,7 @@ namespace Sylius\Behat\Page\Admin\Product;
 use Behat\Mink\Element\NodeElement;
 use Behat\Mink\Session;
 use Sylius\Behat\Behaviour\ChecksCodeImmutability;
-use Sylius\Behat\Context\Ui\Admin\Helper\EditToShowPageSwitcherTrait;
-use Sylius\Behat\Context\Ui\Admin\Helper\ShowPageButtonCheckerTrait;
+use Sylius\Behat\Context\Ui\Admin\Helper\NavigationTrait;
 use Sylius\Behat\Page\Admin\Crud\UpdatePage as BaseUpdatePage;
 use Sylius\Behat\Service\Helper\AutocompleteHelperInterface;
 use Symfony\Component\Routing\RouterInterface;
@@ -25,8 +24,7 @@ use Symfony\Component\Routing\RouterInterface;
 class UpdateSimpleProductPage extends BaseUpdatePage implements UpdateSimpleProductPageInterface
 {
     use ChecksCodeImmutability;
-    use EditToShowPageSwitcherTrait;
-    use ShowPageButtonCheckerTrait;
+    use NavigationTrait;
 
     public function __construct(
         Session $session,
@@ -36,13 +34,11 @@ class UpdateSimpleProductPage extends BaseUpdatePage implements UpdateSimpleProd
         private readonly AutocompleteHelperInterface $autocompleteHelper,
     ) {
         parent::__construct($session, $minkParameters, $router, $routeName);
-
-        $this->defineResourceName();
     }
 
-    protected function defineResourceName(): void
+    protected function getResourceName(): string
     {
-        $this->resourceName = 'product';
+        return 'product';
     }
 
     public function saveChanges(): void
